+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to eww will be listed here, starting at changes since versio
 ### Fixes
 - Fix and refactor nix flake (By: w-lfchen)
 
+### Features
+- Add `:truncate` property to labels, disabled by default (except in cases where truncation would be enabled in version `0.5.0` and before) (By: Rayzeq).
+
 ## [0.6.0] (21.04.2024)
 
 ### Fixes